// src/types/product.ts
export interface Product {
  id: number;
  title: string;
  price: number;
  description: string;
  category: string;
  image: string;
  rating: {
    rate: number;
    count: number;
  };
}

export interface CartItem {
  productId: number;
  title: string;
  image: string;
  price: number;
  quantity: number;
  color: string;
  size: string;
}